Implements shading
Currently implements standard Gouroud shading. No interpolation at the moment, but that's not required by the assignment. Also only computes the shading for one light, but we're not sure if we have to do it for more. Shouldn't be too hard to implement that. As always, comments are added to the code, which should be sufficient for understanding what's happening! Any questions are always welcome :)